uint16_t band a la place de uint8_t band

+++ Core/Arpschuino/libraries/arpschuino/examples/Jeenode_Jeelink/emetteur_JLink_bridge_freq/emetteur_JLink_bridge_freq.ino
@@ -12,7 +12,7 @@ int address (1); //adresse de reception dmx
 int nbre_circuits (66); //nombre de circuit dmx reçu depuis l'adresse
 byte NODEID = 1;  //Adresse RF unique
 #define NETWORKID  212  //adresse du reseau commune à toute les machine
-uint8_t band = RF12_868MHZ; //frequence de l'emeteur RF12
+uint16_t band = RF12_868MHZ; //frequence de l'emeteur RF12
 //RF12_915MHZ America  // RF12_433MHZ
 
 float frequency_setting = 868.00; //beetween 863,00 > 870,00
