+++ Core/Arpschuino/libraries/arpschuino/examples/Jeenode_Jeelink/jNode_to_arpower_8offOn/jNode_to_arpower_8offOn.ino
@@ -12,7 +12,7 @@
 #include <Arpschuino.h>
 byte NODEID = 3;  //Adresse RF unique pour chaque machine
 
-uint8_t band = RF12_868MHZ; //frequence de l'emeteur RF12
+uint16_t band = RF12_868MHZ; //frequence de l'emeteur RF12
 //RF12_915MHZ America  // RF12_433MHZ
 
 #define NETWORKID  212  //adresse du reseau commune à toute les machine
